Fixes using strings for shapes in `MeshRepeatedObjectsTerrainCfg` (#213)
# Description

Meshes such as "box", "cylinder" and "cone" are defined in
`omni.isaac.orbit.terrains.trimesh.utils.py` and can be used in configs
such as
[here](https://github.com/isaac-orbit/orbit/blob/17fc724de7386c919f6f0367980c1d8e9855c446/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/trimesh/mesh_terrains_cfg.py#L187).
In the example, their definition as a string is allowed but as they are
not explicitly imported, the general statement cannot capture the
missing functions. This PR makes sure that all mesh helper functions in
`utils.py` are imported and that the string format can be used.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file