Fixes the RuntimeError when using null command generator (#230)
# Description

The null-command generator class sets the resampling time as
`(-math.inf, math.inf)`. However, on execution, the random ampler does
not like this range and throws a RuntimeError as the operation becomes
out of bounds for float.

To deal with this issue, the MR just overrides all the functions of the
base class and leaves them "empty".

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there