Comments out X11Forwarding bindings in docker-compose.yaml (#348)
# Description

We currently have X11 forwarding enabled by default. This is convenient
for certain users who want graphical displays (including IsaacSim
itself) to be forwarded from within the container. However, it is also
partially dependent upon the configuration of the host system and
requires a valid .Xauthority file which may or may not have been
created. Because this has introduced difficulties for several of our
users, I submit that we comment this portion out be default, until a
solution which allows either a `./container.sh` option, a variable in
`.env`, or both can be used to conveniently enable this option on an
affirmative basis.

If we can agree on this change, I will update the documentation
correspondingly.

## Type of change

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there