Adds preserving of  joint and body indices to the `SceneEntityCfg` (#469)
# Description

The `SceneEntityCfg` always return the `joint_ids` and `body_ids` in the
simulation order even if a different order was specified when passing
the names to the `SceneEntityCfg`. This fix changes the behavior, to
always return the ids in the specified order by introducing the flag
`simulation_order` that is true per default for everything within orbit
but false for the `SceneEntityCfg`.

Fixes #461 

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: Pascal Roth <57946385+pascal-roth@users.noreply.github.com>
Signed-off-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>
Co-authored-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>
Co-authored-by: Mayank Mittal <mittalma@leggedrobotics.com>