Fixes `Sb3VecEnvWrapper` to clear buffer on reset (#974)
# Description

In previous version of the SB3 environment wrapper, the episode buffer
was not cleared when `env.reset` was called. This led to an
overestimation of the number of time-steps and rewards in subsequent
episodes, as reflected in the `infos` returned by `env.steps`. This
commit aims to address this.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ]  I have made corresponding changes to the documentation
- [x]  My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there