Fixes scaling of default ground plane (#1133)
# Description

A typo in the code was preventing the `size` parameter in GroundPlaneCfg
from being parsed and applied to the default ground plane. In addition,
some errors are being thrown with the kit command use to apply the
scale. This change fixes the typo and switches to use isaac sim core
APIs for applying scale.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there